python_engine/requirements.txt
bidict==0.21.2
cachelib==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dnspython==1.16.0
et-xmlfile==1.1.0
eventlet==0.30.2
Flask==1.1.2
Flask-CLI==0.4.0
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SocketIO==5.0.1
greenlet==1.0.0
idna==2.10
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==2.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
natsort==8.0.2
numpy==1.20.2
opencv-python==4.5.1.48
openpyxl==3.0.7
pandas==1.3.0
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.17.1
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
pywin32==300
requests==2.25.1
scipy==1.7.0
six==1.15.0
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1

packaging~=21.3